old-tests/submission/Infraware/Forms/contents/Forms/select_selectedoptions.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Forms</title>
  <script src="../../resources/testharness.js"></script>
  <script src="../../resources/testharnessreport.js"></script>
 </head>
 <body>
	<p>
		<h3>select_selectedOptions</h3>
	</p>
	
	<hr>

	<div id="log"></div>

	<form method="post"
		  enctype="application/x-www-form-urlencoded"
		  action=""
		  id="input_form">
		<p>
		 <label for="unittype">Select unit type:</label>
		 <select id="unittype" name="unittype" multiple>
		  <option value="1" selected> Miner </option>
		  <option value="2" selected> Puffer </option>
		  <option value="3"> Snipey </option>
		  <option value="4"> Max </option>
		  <option value="5" selected> Firebot </option>
		 </select>
		</p>	
	</form>

	<script>
		
		var select = document.getElementById("unittype");
		var selectedoptions = select.selectedOptions;

		test(function() {
			assert_equals(selectedoptions.length, 3, "selectedOptions attribute is not correct.");
		});

	</script>

 </body>
</html>
